# src/etapi/etapi.openapi.yaml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: ETAPI
  description: External Trilium API
  contact:
    name: zadam
    email: zadam.apps@gmail.com
    url: https://github.com/zadam/trilium
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:37740/etapi
  - url: http://localhost:8080/etapi
security:
  - EtapiTokenAuth: []
paths:
  /create-note:
    post:
      description: Create a note and place it into the note tree
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteDef'
      responses:
        '200':
          description: note created
          content:
            application/json:
              schema:
                properties:
                  note:
                    $ref: '#/components/schemas/Note'
                    description: Created note
                  branch:
                    $ref: '#/components/schemas/Branch'
                    description: Created branch
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /notes/{noteId}:
    parameters:
      - name: noteId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/EntityId'
    get:
      description: Returns a note identified by its ID
      operationId: getNoteById
      responses:
        '200':
          description: note response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      description: patch a note identified by the noteId with changes in the body
      operationId: patchNoteById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        '200':
          description: update note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: deletes a single note based on the noteId supplied
      operationId: deleteNoteById
      responses:
        '204':
          description: note deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /branches/{branchId}:
    parameters:
      - name: branchId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/EntityId'
    get:
      description: Returns a branch identified by its ID
      operationId: getBranchById
      responses:
        '200':
          description: branch response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      description: create a branch (clone a note to a different location in the tree)
      operationId: postBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Branch'
      responses:
        '200':
          description: update branch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      description: patch a branch identified by the branchId with changes in the body
      operationId: patchBranchById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Branch'
      responses:
        '200':
          description: update branch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: >
        deletes a branch based on the branchId supplied. If this is the last branch of the (child) note, 
        then the note is deleted as well.
      operationId: deleteBranchById
      responses:
        '204':
          description: branch deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /inbox/{date}:
    get:
      description: > 
        returns an "inbox" note, into which note can be created. Date will be used depending on whether the inbox
        is a fixed note (identified with #inbox label) or a day note in a journal.
      operationId: getInboxNote
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string 
            format: date
          example: 2022-02-22
      responses:
        '200':
          description: inbox note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendar/days/{date}:
    get:
      description: returns a day note for a given date. Gets created if doesn't exist.
      operationId: getDayNote
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          example: 2022-02-22
      responses:
        '200':
          description: day note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendar/weeks/{date}:
    get:
      description: returns a week note for a given date. Gets created if doesn't exist.
      operationId: getWeekNote
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          example: 2022-02-22
      responses:
        '200':
          description: week note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendar/months/{month}:
    get:
      description: returns a week note for a given date. Gets created if doesn't exist.
      operationId: getMonthNote
      parameters:
        - name: month
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9]{4}-[0-9]{2}'
          example: 2022-02
      responses:
        '200':
          description: month note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /calendar/years/{year}:
    get:
      description: returns a week note for a given date. Gets created if doesn't exist.
      operationId: getYearNote
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9]{4}-[0-9]{2}'
          example: 2022-02
      responses:
        '200':
          description: year note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    EtapiTokenAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    CreateNoteDef:
      type: object
      required:
        - parentNoteId
        - type
        - title
        - content
      properties:
        parentNoteId:
          $ref: '#/components/schemas/EntityId'
          description: Note ID of the parent note in the tree
        type:
          type: string
          enum: 
            - text
            - code
            - file
            - image
            - search
            - book
            - relation-map
            - render
        mime:
          type: string
          description: this needs to be specified only for note types 'code', 'file', 'image'.
          example: application/json
        title:
          type: string
        content:
          type: string
        notePosition:
          type: integer
          description: > 
            Position of the note in the parent. Normal ordering is 10, 20, 30 ... 
            So if you want to create a note on the first position, use e.g. 5, for second position 15, for last e.g. 1000000
        prefix:
          type: string
          description: >
            Prefix is branch (placement) specific title prefix for the note. 
            Let's say you have your note placed into two different places in the tree, 
            but you want to change the title a bit in one of the placements. For this you can use prefix.
        noteId:
          $ref: '#/components/schemas/EntityId'
          description: DON'T specify unless you want to force a specific noteId
        branchId:
          $ref: '#/components/schemas/EntityId'
          description: DON'T specify unless you want to force a specific branchId
    Note:
      type: object
      properties:
        noteId:
          $ref: '#/components/schemas/EntityId'
          readOnly: true
        title:
          type: string
        type:
          type: string
          enum: [text, code, book, image, file, mermaid, relation-map, render, search, note-map]
        mime:
          type: string
        isProtected:
          type: boolean
          readOnly: true
        attributes:
          $ref: '#/components/schemas/AttributeList'
          readOnly: true
        parentNoteIds:
          $ref: '#/components/schemas/EntityIdList'
          readOnly: true
        childNoteIds:
          $ref: '#/components/schemas/EntityIdList'
          readOnly: true
        parentBranchIds:
          $ref: '#/components/schemas/EntityIdList'
          readOnly: true
        childBranchIds:
          $ref: '#/components/schemas/EntityIdList'
          readOnly: true
        dateCreated:
          $ref: '#/components/schemas/LocalDateTime'
          readOnly: true
        dateModified:
          $ref: '#/components/schemas/LocalDateTime'
          readOnly: true
        utcDateCreated:
          $ref: '#/components/schemas/UtcDateTime'
          readOnly: true
        utcDateModified:
          $ref: '#/components/schemas/UtcDateTime'
          readOnly: true
    Branch:
      type: object
      description: Branch places the note into the tree, it represents the relationship between a parent note and child note
      required:
        - noteId
        - parentNoteId
      properties:
        branchId:
          $ref: '#/components/schemas/EntityId'
          readOnly: true
        noteId:
          $ref: '#/components/schemas/EntityId'
          readOnly: true
          description: identifies the child note
        parentNoteId:
          $ref: '#/components/schemas/EntityId'
          readOnly: true
          description: identifies the parent note
        prefix:
          type: string
        notePosition:
          type: integer
          format: int32
        isExanded:
          type: boolean
        utcDateModified:
          $ref: '#/components/schemas/UtcDateTime'
          readOnly: true
...
    AttributeList:
      type: array
      items:
        $ref: '#/components/schemas/Attribute'
...
    EntityId:
      type: string
      pattern: '[a-zA-Z0-9]{4,12}'
      example: evnnmvHTCgIn
    EntityIdList:
      type: array
      items:
        $ref: '#/components/schemas/EntityId'
    LocalDateTime:
      type: string
      pattern: '[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3}\+[0-9]{4}'
      example: 2021-12-31 20:18:11.939+0100
    UtcDateTime:
      type: string
      pattern: '[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3}Z'
      example: 2021-12-31 19:18:11.939Z
    Error:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status, identical to the one given in HTTP response
          example: 400
        code:
          type: string
          description: stable string constant
          example: NOTE_IS_PROTECTED
        message:
          type: string
          description: Human readable error, potentially with more details,
          example: Note 'evnnmvHTCgIn' is protected and cannot be modified through ETAPI